Fix deadlock when sending to gone client
The problem was grabbing an RLock while trying
to send to the sendCh, whilst there could be
no one on the receiving side anymore and the
channel send getting blocked forever. That would
block forever while holding an RLock, preventing
anyone else from changing the subscriptions state,
including the deactivation of that exact same sub.

For that, added a done channel to the subscription
struct and a lock to safely guarantee that we don't
try to send on a potentially closed channel.