Use a simple Mutex instead of RWMutex
We are only supposed to grab the RLock from a
single goroutine anyway, so it wasn't really
necessary to use an RLock there.

Additionally, by using a simple Mutex, we also ensure
that even if some other code error is done, the resetConn
method will still be executed only once at the same time,
as it peforms a mutation on the current connection that
could perform badly in case of a race.