Add cache library
This library has been written for a specific service, but it has proven that it
is good enough to be shared and, thus, share it we will.

We have already started to test this locally in the service where the library
comes from, since it is the easiest to update with this.