Fix memory leak :scream:
We were using Tick which is meant for ticks
that never stop, not only a single tick,
and since we don't have access to the underlying
ticker, we can never dispose of it.

So every time we pinged, we created an object that
was never collected by the garbage collector :this-is-fine-fire:

Fix it by using time.After instead of time.Tick.
Also replace the ping ticker with a time.Tick, since
we don't need access to the ticker directly (we'll never
dispose of it either, but in case it's fine since we never
stop using it).