Remove cyclic reference between conn and loop state
We should keep the "common communication layer" in the
subConn "parent" struct, with the inner loop state
referencing it for whatever it needs, while the "subConn"
itself doesn't even need to keep a reference to its inner
"main loop state", as it will be just happening forever in
background. The only thing we should ever have is like a
way to cancel the loop, like a done channel, but that is
not necessary for our current interface so I'm keeping it
only one-way for now.