Fix Redis connection concurrent usage (#43)
* Subscribe through main loop to avoid concurrency

* Do unsubscription on main loop as well

* Remove mutex :sparkles:

All operations will be performed on the main loop
so we don't need to have a mutex anymore :sunglasses:

* Be smarter about bad input and errors

Avoid dropping valid requests by ensuring we have
at least one retry on a new, surely good connection
before just going on without performing the requested
subscription.

We should have an alarm on that subscribe_error log,
since it means a subscription which we have silently
dropped while have returned a successful response to
our caller.

* Centralize error handling in the main loop

* Create helper function for only the main iteration

This way the looping and error handling/recovering
is kept in one place and only the denser business
logic is in the mainIteration func.

* Avoid the named returned values

Much clearer without them TBH

* Move message handling to between ping and pong timeout

Makes it more obvious where the pong is expected to come
from.

* Use fixed err instances for known errors

* Always log when recovering connection

Will allow to know when we recover from a
psubscribe attempt, since in that case we
don't log before doing it.

* Avoid possibly sending a PUnsubscribe with no channels

* Add a timeout on subscribe in case chan is full

We do not want to create a lot of hanged go-routines
in case the main loop gets blocked for some reason.

* Fix recovering of connection before retry

We were using the "recover" name which already has
a default meaning in the go language and tho didn't
generate any errors :harold-pain:

Fix it by passing the function to the iteration
function, and avoid the bug in the future by renaming
it to a more specific recoverConn

* Fix memory leak :scream:

We were using Tick which is meant for ticks
that never stop, not only a single tick,
and since we don't have access to the underlying
ticker, we can never dispose of it.

So every time we pinged, we created an object that
was never collected by the garbage collector :this-is-fine-fire:

Fix it by using time.After instead of time.Tick.
Also replace the ping ticker with a time.Tick, since
we don't need access to the ticker directly (we'll never
dispose of it either, but in case it's fine since we never
stop using it).

* Improve error message

* Create separate pubSubLoopState to hold the... state

I'm not done yet, but I think this is a good first step :)

* Remove cyclic reference between conn and loop state

We should keep the "common communication layer" in the
subConn "parent" struct, with the inner loop state
referencing it for whatever it needs, while the "subConn"
itself doesn't even need to keep a reference to its inner
"main loop state", as it will be just happening forever in
background. The only thing we should ever have is like a
way to cancel the loop, like a done channel, but that is
not necessary for our current interface so I'm keeping it
only one-way for now.

* Move all the state updating to the resetConn method

We shouldn't be calling a reccover/reset method and
then need to do more stuff to reach a valid state.
That method should do it itself and we even avoid some
code repetition as well as keep the logic in a single place.

* Use context instead of a manual cancellation chan

OK now I'm done :)

* Decrease timeouts and increase buffer sizes

Create some const for the buffer size so we
don't repeat it everywhere.

Also add a comment for why we don't have a
timeout for *unsubscribing* right now.

* Improve management of redis msg channel

 - Rename cancel/close func to closeMsgChan
 - Actually close the msgChan when that one is triggered
 - Drain the previous channel buffer when creating a new one

* Separate the receiving and sending to channel

Avoid making it seem like it is a single select statment,
when in actual execution the stack will block inside the
Receive function and only after something is received that
we would get into the select statment. Keeping it separate
makes it a bit clearer now.

* Handle redis.Subscription messages from Redis