Add Broker&BrokerPool implementation from Colossus
Back in the day, we had this implementation before
the actual subscriptions to Redis. At some point I
realized the all pubsub logic we have in the
redis/pubsub.go client made all this brokering logic
unnecessary.

I still think this is some pretty good code though,
and had it sitting in my local go-io repository for
quite a while now.

So let's just commit and open source it, and maybe
it can be useful for anyone in the future :)