Bring back the lock to Subscribe function
Realized there could still be a race condition
between the unsubscribe and the reading of the
map so oh well.

It didn't even need to be a sync.Map after all
if we are locking when accessing it anyway, but
I'll keep it since it allows it to always be
an RLock instead of needing a WLock when making
changes to it like adding a new subscriber.