Add Broker&BrokerPool implementation from Colossus (#45)
* Add Broker&BrokerPool implementation from Colossus

Back in the day, we had this implementation before
the actual subscriptions to Redis. At some point I
realized the all pubsub logic we have in the
redis/pubsub.go client made all this brokering logic
unnecessary.

I still think this is some pretty good code though,
and had it sitting in my local go-io repository for
quite a while now.

So let's just commit and open source it, and maybe
it can be useful for anyone in the future :)

* Use sync.Map on broker pool

* Bring back the lock to Subscribe function

Realized there could still be a race condition
between the unsubscribe and the reading of the
map so oh well.

It didn't even need to be a sync.Map after all
if we are locking when accessing it anyway, but
I'll keep it since it allows it to always be
an RLock instead of needing a WLock when making
changes to it like adding a new subscriber.